backend/requirements.txt
fastapi==0.109.0
uvicorn==0.27.0
pydantic==2.5.3
pydantic-settings==2.1.0
openai==1.12.0
chromadb==0.4.22
sqlalchemy==2.0.25
aiosqlite==0.19.0
python-multipart==0.0.6
pytesseract==0.3.10
pillow==10.2.0
python-docx==1.1.0
openpyxl==3.1.2
pandas==2.2.0
PyPDF2==3.0.1
tabulate==0.9.0
watchdog==4.0.0
python-dotenv==1.0.1
httpx==0.26.0
email-validator==2.1.0.post1
jinja2==3.1.3
